Add cell variants for generic load/store
In the Rust type system these are fundamentally different from normal
references, in that they may be aliased by other pointers. However, the
load and store instructions do not care about this as they operate on
memory and these intrinsics do not enforce aliasing.

This even applies to storeu2 which are macros around a set of two stores
of size u128.