Add non-temporal store abstraction
Note that non-temporal loads in x86 do not need an abstraction in such a
manner. They do not contradict Rust's memory model. As all x86 loads are
ordered in a particular processor even non-temporal loads that bypass
the cache obey the necessary order.

In contrast, non-temporal stores will write values in a manner that
would *not* be observed by following loads. To allow stores according to
the memory model we build an abstraction that simply makes such loads
impossible. A scope ensures that within it all accesses are stores and
that all access following the scope has ran into an sfence instruction
before the load, thus obeying the established order relationships.